requirements.txt
streamlit
requests
beautifulsoup4
Pillow
urllib3